refactor : service에서 input, output을 view package로 바꾸고 나머지는 domain package로 이동
