+++ src/test/kotlin/next/step/racing/domain/StepCountTest.kt
@@ -4,7 +4,7 @@ import io.kotest.assertions.throwables.shouldThrow
 import io.kotest.core.spec.style.DescribeSpec
 import io.kotest.datatest.withData
 import io.kotest.matchers.shouldBe
-import next.step.racing.service.to.StepCount
+import next.step.racing.domain.racing.StepCount
 
 class StepCountTest : DescribeSpec({
